+++ doc/release-notes.rst
@@ -4,7 +4,7 @@ History of Changes
 
 .. Upcoming Version
 .. ----------------
-* adds the following new GEM datasets: GCPT, GSPT, GGtPT, GWPT and updates GGPT. 
+* updates contributing docs for consistency with latest powerplantmatching version. 
 
 
 Version 0.5.6 (13.02.2023)