+++ doc/release-notes.rst
@@ -14,7 +14,7 @@ History of Changes
   * `GNPT`, nuclear power plant
   * `GSPT`, solar power plant
   * `GWPT`, wind power plant 
-
+* Replace unstable FIAS and OPSD download links by new provider
 
 Version 0.5.6 (13.02.2023)
 -------------------------