+++ doc/release-notes.rst
@@ -6,6 +6,19 @@ History of Changes
 .. ----------------
 **New Features**
 
+
+
+Version 0.5.7 (30.05.2023)
+-------------------------
+
+
+**Bug fixes**
+
+* Replace unstable FIAS and OPSD download links by new provider
+* Replace ValueError when no matches found to logged warning.
+
+**Other Changes**
+
 * adds the following new Global Energy Monitor data including:
   * `GBPT`, bioenergy power plant
   * `GGPT`, gas power plant update
@@ -15,8 +28,7 @@ History of Changes
   * `GSPT`, solar power plant
   * `GWPT`, wind power plant 
   * `GHPT`, hydro power plant
-* Replace unstable FIAS and OPSD download links by new provider
-* Replace ValueError when no matches found to logged warning.
+
 
 Version 0.5.6 (13.02.2023)
 -------------------------